backend/requirements.txt
Django
djangorestframework
django-cors-headers
dj-rest-auth
django-allauth
python-dotenv
PyPDF2
langchain
langchain_community
openai
sentence-transformers
chromadb
HuggingFace_hub
Pillow
langchain_openai
jwt
# For Google OAuth2 integration
oauthlib
requests-oauthlib
pypdf
boto3
django-storages
mysqlclient
docker
faiss-cpu
social-aut-app-django
djangorestframework
# chromadb
